Added a test project of a collision engine
Pulled an all-nighter writing this from scratch because I was bored and
I felt like we need some actual code written.  This might be a bit much
to wrap your head around at first, feel free to ask me questions in
class/on Facebook.

The idea of this program is to test the CollisionManager class.  I found
a good explanation for how to construct it, and refactored the code to
fit with the rest of the project.  This code has native support for
bounding boxes of any polygon shape and can supply us with a way to have
that "sliding wall" detection we want, where running into a wall at an
angle gives a friction effect - all it needs is another method or two.

The way the program works now is you can control the Boo with the arrow
keys, and if it overlaps with a Dry Bones, the bounding box on the Dry
Bones will turn red.  This is to show how this collision detection can
differentiate who is doing the colliding, and which entity to perform an
action on.  Note how the collision box is separate from the sprite, this
could give us an option for pseudo-top-down like Garrett wants if we can
figure out the details since the sprites themselves are not doing the
colliding.  It also lets us give each entity its own collision shape
independent of the sprite shape.