Grid Checking
Added a new GridCheck class
Changed collision detection to use the GridCheck class

The game will now only perform collision checks on entities that are in
the same grid tile(s) as other entities.  In a theoretical situation
where there are 100 entities on screen, a regular update would need
10,000 collision checks, where this could only need a couple hundred or
so.  Hoping to see if this could pave the way to Entity/Wall checks.

New code is mostly uncommented.