3/4 View
Since this version of the engine is best suited to a 3/4 View
perspective (i.e. Legend of Zelda, Pokemon, Earthbound, etc) and we have
decided on a pure top-down, I'm pushing it to a separate branch since
the engine will have to work differently and will need to be rewritten.
This will be a reference branch.