Basic Player Sprite/Environment Sheets! :+1:
BasicRoomTemplate.jpg is a close representation to how perspective is
used in old Zelda dungeons. It's a good idea of what I personally had in
mind for our layout but we can always discuss it more later. The room is
pretty big in the example so it would probably be applicable to use the
view class for that dimension. Once again, they are only test sprites.
I'll make better looking borders later on.
Both png's added are 320x320 because I plan to add more stuff later
(like better sprites and animation). I don't know how much this messes
with the current tiling system but feel free to change them if need be.
If anyone figures out how to arrange the sprites like the .jpg file, you
get a gold star.

Finally there is the player sprite sheet. Thought we should have at
least some representation of our character. I actually like the colors
but I'll be definitely changing things in the future. Both the player
and gauntlet are 32x32, so hopefully they will suffice for now.