Added My Part To Game Design Template
I thought it would be easier just for everyone to modify this document
and add their part so one person isn't paired up with batching all of
the stuff together. I didn't stray away from using the table of contents
as is so nothing has to be changed there. Currently working on some
enemy sprites and stuff so I'll hopefully get that out soon.