Some Music and Splash Screen
Still working on the sprites, but I thought I should just upload some
other stuff in the meantime. NoCombat music has been changed to 4/4 time
to compensate for the combat music. I tried making the combat music in
3/4 time like the original nocombat music but failed terribly. Both
should be streamable if we wish to do that in the future, although one
will have to fade out while the other fades in order to be
comprehensive. Remember to adjust your volume since it doesn't just fade
in, it might be a little loud.
Also I made a quick splash screen. I'm a little mad about it because it
was originally in 60 frames and looked great, but avi turns computers
into potatoes. You might have to replay it a couple times since it's
only 2 seconds.