Some Clean Up, Leveling Up, and Release Build
I was having trouble with the Debug build of the project. The executable
was completely gone. This seems to happen when I mess around with XACT,
no idea why. However, the release version still works. If there
objections to working with the release build, we can always revert this
build. Or someone can figure out the mystery as to why the Debug build
is missing its executable file.
Added a level up sound. There is supposed to be text displayed
indicating a level up, but I'm having trouble getting it to play once
and then vanish. For now, it kind of twitches constantly in the left up
corner for example purposes. There is also an mp4 to show what I want it
to do. Of course, turning it into a sprite sheet meant losing some
quality, but it's still legible.