Room Workin'
Changed the fountain room to have a wider teleport to compensate for the
mausoleum room. Of course I had to take out some grass tiles and now it
looks worse but at least you can't go outside of the teleport anymore.
Well, you can still go around with the other entrances, but whatever.